managers linux, php et mysql fonctionnels
testés sans les commandes shell sous WINDOWS

+++ includes/PHPmanager.class.php
@@ -18,30 +18,31 @@ class PHPmanager {
 
         return "[$identifiant]
 
-        user = $identifiant
-        group = $identifiant
-        
-        listen = ".PHP_SOCKET_PATH."$identifiant-fpm.sock
-        
-        listen.owner = www-data
-        listen.group = www-data
-        
-        pm = dynamic
-        
-        pm.max_children = 5
-        
-        pm.start_servers = 2
-        
-        pm.min_spare_servers = 1
-        
-        pm.max_spare_servers = 3
-        
-        chroot = ".PHP_PATH."$identifiant/www/
-        chdir = /
-        
-        php_admin_value[open_basedir]   = /
-        php_admin_value[sys_temp_dir]   = /tmp/
-        php_admin_value[upload_tmp_dir] = /tmp/";
+user = $identifiant
+group = $identifiant
+
+listen = ".PHP_SOCKET_PATH."$identifiant-fpm.sock
+
+listen.owner = www-data
+listen.group = www-data
+
+pm = dynamic
+
+pm.max_children = 5
+
+pm.start_servers = 2
+
+pm.min_spare_servers = 1
+
+pm.max_spare_servers = 3
+
+chroot = ".PHP_PATH."$identifiant/www/
+chdir = /
+
+php_admin_value[open_basedir]   = /
+php_admin_value[sys_temp_dir]   = /tmp/
+php_admin_value[upload_tmp_dir] = /tmp/
+        ";
     }
 
     /* Ecrit la conf */
@@ -59,7 +60,7 @@ class PHPmanager {
 
         file_put_contents(
             PHP_PATH . $identifiant. ".conf",
-            $this->genererConf($identifiant);
+            $this->genererConf($identifiant)
         );
     }
 
@@ -73,6 +74,11 @@ class PHPmanager {
         if(!is_dir(PHP_PATH) || !is_writable(PHP_PATH))
         {
             throw new PHPManagerException("Le chemin vers les pools PHP n'existe pas ou n'est pas accessible en écriture");
+        }
+
+        if(!file_exists(PHP_PATH . $identifiant . ".conf"))
+        {
+            throw new PHPmanagerException("Le fichier de conf associé à cet identifiant n'existe pas");
         }
 
         unlink(PHP_PATH . $identifiant . ".conf");
